Refactor: community·mypage 라우트 폴더명을 일관되게 정리
