+++ src/app/community/[id]/edit/page.tsx
@@ -1,0 +1,5 @@
+import EditPostForm from "@/features/community/postEdit/EditPostForm";
+
+export default function EditPage() {
+  return <EditPostForm />;
+}
