--- src/app/community/edit/[id]/page.tsx
@@ -1,5 +1,0 @@
-import EditPostForm from "@/features/community/post-edit/EditPostForm";
-
-export default function EditPage() {
-  return <EditPostForm />;
-}
