+++ src/app/mypage/myPosts/[category]/page.tsx
