--- src/app/mypage/page.tsx
@@ -1,3 +1,0 @@
-export default function MyPage() {
-  return <div> 페이지</div>;
-}
